Memoized carousel for performance
