--- bloom-and-vine/src/Components/gallery-carousel/Components/galleryImage/galleryImage.css
@@ -1,27 +1,0 @@
-/* .gallery-image {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-} */
-
-.gallery-primary-image {
-    height: 448px;
-    object-fit: cover;
-    min-width: 25vw;
-    max-width: 25vw;
-}
-
-.gallery-secondary-image {
-    opacity: 0.5;
-    width: 16vw;
-    height: 375px;
-    object-fit: cover
-}
-
-.gallery-tertiary-image {
-    position: relative;
-    opacity: 0.5;
-    width: 14vw;
-    height: 321px;
-    object-fit: cover
-}
