+++ bloom-and-vine/src/Components/gallery-carousel/Components/galleryInfo/galleryInfo.css
@@ -3,6 +3,9 @@
     flex-direction: column;
     align-items: center;
     justify-content: center;
+
+    font-style: normal;
+    text-align: center;
 }
 
 .upper {
@@ -15,24 +18,18 @@
 
 .gallery-title {
     font-family: 'Italiana';
-    font-style: normal;
     font-size: 3.56em;
     font-weight: 400;
-    text-align: center;
 }
 
 .gallery-date {
     font-family: 'Cormorant Garamond';
-    font-style: normal;
     font-size: 1.57em;
     font-weight: lighter;
-    text-align: center;
 }
 
 .gallery-desc {
     font-family: 'Cormorant Garamond';
-    font-style: normal;
     font-size: 1.57em;
     font-weight: 500;
-    text-align: center;
 }
