+++ bloom-and-vine/src/Components/gallery-carousel/Components/gallerySlider/gallerySlider.js
@@ -1,5 +1,5 @@
-import { useState, useEffect } from 'react';
-import GalleryImage from '../galleryImage/galleryImage';
+import { useState, useEffect, useCallback } from 'react';
+import GalleryImage from './components/galleryImage/galleryImage';
 import BackArrow from '../../../../Images/back_arrow.png'
 import ForwardArrow from '../../../../Images/forward_arrow.png'
 import './gallerySlider.css'
@@ -15,46 +15,47 @@ function getGalleryDots(total, setCurImgDot, level) {
         return dot_display;
     }
     
-function setCurrentDot(index, images, level) {
-    const curIndex = index + 1 <= images.length ? index + 1 : 1;
-    const inactiveDots = document.querySelectorAll(`.gallery-dot-${level}:not(:nth-child(${curIndex}))`)
-    const curDot = document.querySelector(`.gallery-dot-${level}:nth-child(${curIndex})`)
     
-    for (let i = 0; i < inactiveDots.length; i++) {
-        inactiveDots[i].style.backgroundColor = "#D9D9D9";
-    }
-    
-    curDot.style.backgroundColor = "#A28598";
-}
-
 export default function GallerySlider (props) {
     const [primaryImgIndex, setPrimaryImgIndex] = useState(0);
     const images = props.data.images;
     const level = props.level;
     
     const galleryDots = getGalleryDots(images.length, setPrimaryImgIndex, level);
-
-    function onGalleryClickHandler(e, button) {
+    
+    const onGalleryClickHandlerMemoized = useCallback((e, button) => {
         e.preventDefault();
-
+        
         if (button === 'left') {
             primaryImgIndex - 1 >= 0 ? setPrimaryImgIndex(prevIndex => prevIndex - 1) : setPrimaryImgIndex(images.length - 1);
         }
-    
+        
         if (button === 'right') {
             primaryImgIndex + 1 < images.length ? setPrimaryImgIndex(prevIndex => prevIndex + 1) : setPrimaryImgIndex(0);
         }
-    }
+    }, [primaryImgIndex]);
+
+    const setCurrentDotMemoized = useCallback((index, images, level) => {
+        const curIndex = index + 1 <= images.length ? index + 1 : 1;
+        const inactiveDots = document.querySelectorAll(`.gallery-dot-${level}:not(:nth-child(${curIndex}))`)
+        const curDot = document.querySelector(`.gallery-dot-${level}:nth-child(${curIndex})`)
+        
+        for (let i = 0; i < inactiveDots.length; i++) {
+            inactiveDots[i].style.backgroundColor = "#D9D9D9";
+        }
+        
+        curDot.style.backgroundColor = "#A28598";
+    }, [primaryImgIndex]);
 
     useEffect(() => {
-        setCurrentDot(primaryImgIndex, images, level);
+        setCurrentDotMemoized(primaryImgIndex, images, level);
     }, [primaryImgIndex])
     
     
     return (
         <div className='gallery-slider-container'>
             <div className='gallery-slider'>
-                <button className='gallery-event-button back-button' onClick={e => onGalleryClickHandler(e, 'left')}>
+                <button className='gallery-event-button back-button' onClick={e => onGalleryClickHandlerMemoized(e, 'left')}>
                     <img className='button-image' src={BackArrow} />
                 </button>
                 <GalleryImage 
@@ -86,7 +87,7 @@ export default function GallerySlider (props) {
                         altText={images[primaryImgIndex + 2 > images.length - 1 ? primaryImgIndex + 2 - images.length : primaryImgIndex + 2].altText} 
                         imgClass={'gallery-tertiary-image'} 
                     />
-                <button className='gallery-event-button forward-button' onClick={e => onGalleryClickHandler(e, 'right')}>
+                <button className='gallery-event-button forward-button' onClick={e => onGalleryClickHandlerMemoized(e, 'right')}>
                     <img className='button-image' src={ForwardArrow} />
                 </button>
             </div>
