+++ bloom-and-vine/src/Components/gallery-display-mobile/galleryDisplayMobile.css
@@ -1,32 +1,36 @@
-#GallMobHeader{
-font-family: 'Italiana';
-font-style: normal;
-font-weight: 400;
-font-size: 40px;
-line-height: 47px;
-color: #000000;
-text-align: left;
-margin-left: 21px;
-margin-bottom: 24px;
-margin-top: 24px;
+#GallMobHeader {
+    font-family: 'Italiana';
+    font-style: normal;
+    font-weight: 400;
+    font-size: 40px;
+    line-height: 47px;
+    color: #000000;
+    text-align: left;
+    margin-left: 21px;
+    margin-bottom: 24px;
+    margin-top: 24px;
 }
-#categoryList{
+
+#categoryList {
     display: flex;
     flex-direction: column;
     align-items: center;
     justify-content: center;
 }
-.categoryBox{
+
+.categoryBox {
     display: flex;
     align-items: center;
     justify-content: center;
     margin-bottom: 20px;
 }
-.categoryImg{
+
+.categoryImg {
     width: 100vw;
     filter: brightness(60%)
 }
-.categoryText{
+
+.categoryText {
     position: absolute;
 
     font-family: 'Cormorant Garamond';
